Updated components for breakpoint 640px
