+++ src/components/page-components/our-projects/components/ServiceItemSmall.vue
@@ -1,5 +1,7 @@
 <template>
-  <div class="col-span-2 md:col-span-1 md:px-4">
+  <div
+    class="col-span-4 sm:col-span-2 md:col-span-1 md:px-4 mobile:text-center"
+  >
     <div class="mb-5">
       <i
         class="bg-gradient-to-br from-pink to-purple bg-clip-text text-4xl text-transparent"
